fix problems after merge
