+++ pytorch_lightning/hypertuner/hypertuner.py
@@ -1,4 +1,4 @@
-from typing import Union, Optional, List, Dict, Tuple, Iterable, Any
+from typing import Union, Optional, List
 
 from torch.utils.data import DataLoader
 
@@ -45,6 +45,9 @@ class HyperTuner(HyperTunerLRFinderMixin, HyperTunerBatchScalerMixin):
         # For checking dependency
         self._scale_batch_size_called = False
         self._lr_find_called = False
+        
+        self._prepare_data_called = False
+        self._model_setup_called = False
 
         rank_zero_warn('HyperTuner class is `experimental`, meaning that some of'
                        ' its functionality is under tested and its interface may'
