+++ pytorch_lightning/trainer/training_tricks.py
@@ -5,6 +5,7 @@ import torch
 from torch import Tensor
 
 from pytorch_lightning import _logger as log
+from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.callbacks import GradientAccumulationScheduler
 
 EPSILON = 1e-6
